Caching the inverse of a matrix
First function  makeCacheMatrix creates a list containing functions to
1. set the value of the matrix
2. get the value of the matrix
3. set the value of inverse of the matrix 4. get the value of inverse of
the matrix
Second function returns the inverse of the matrix.
It first checks if the matrix has  been changed.
If so, it stops, gets the message and skips the computation.
If not,  it checks if the inverse has already been computed.
If so, it gets the result and skips the  computation.
If not, it computes the inverse, sets the value in the cache.
This function assumes that the matrix is always invertible.